requirements.txt
django == 4.2.3
django-autoslug == 1.9.9
